Added a SHA1 for stress test and moved to end of test suite list.
The stress tests are intended to provide a way to test the limits of
CellBlender in various ways. The current approach is to provide a single
stress level parameter that will then be used to create and
appropriately "stressful" model. The number of things (objects,
molecules, parameters, reactions, plots, etc) represented by a
particular stress level may not match that stress level (or even be
linear with that stress level). But the intention is that a stress level
of 100 would be a "normal" number with higher numbers representing more
stressful tests. Right now, the stress level matches the number of
objects that are created (since that's the only stress test).

Note also, that the SHA1 that's stored for stress tests will only match
the setting of 100. So the test suite will report an error for other
stress test settings. That should be fine since the primary purpose of
these tests is to measure performance with correctness already being
tested in other tests.