Began working on pipe prototypes similar to RenderMan's RIB
The RenderMan Interface Bytestream provides a way for any program
(written in any language) to communicate with RenderMan. This commit
contains experiments with that kind of interface to see how it might
work for MCell. Both a C and C++ version have been created, but work
will likely focus on the C++ version first.