Implemented a timer callback and a molecule creation callback
CellBlender currently has a subdirectory named "libMCell" which contains
a few sandbox prototypes in C, C++, and Python for experimenting with
CellBlender interfaces to different simulations.

The code is very simple, and it does not produce numerically correct
results. However, it does interact properly with CellBlender and it is a
good testing ground for new ideas.

This commit introduces two C++ callback event class prototypes:

  MCellTimerEvent
  MCellMoleculeCreationEvent

These can be appended to the simulation's timer and mol creation event
handlers lists. The simulation will then call them for each iteration
and whenever molecules are created (respectively). The program named
mcell_simple_count.cpp illustrates how these classes can be overloaded
(customized) and appended to the appropriate callback lists.

mcell_simple_count.cpp is very short (136 lines) and it's intended to be
a preview of how we might use libMCell from C++. Since we don't have
variable parameters in C++, this version simply exposes all of its
member data as public so it can be set directly by the user. This isn't
the best style, but it's very efficient and works well at this scale.