Quick fix for KeyError: 'bpy_prop_collection[key]: key "p15" not found'
This is a quick fix for the KeyError problem reported in the parameter
system earlier this month.

The problem manifests itself in circumstances where the CellBlender
properties are rebuilt from a data model. The traceback will show:

   KeyError: 'bpy_prop_collection[key]: key "p#" not found'

where the "p#" may take on different numbers (15 is common).

The problem stems from the automatic creation of Blender's "RNA"
properties when new property groups are created. The new CellBlender
parameter system must create dynamic ID properties when these "RNA"
properties are created. This works well for items that are dynamically
created and destroyed (molecules, reactions, etc). However, there are
some CellBlender properties that are static (time step, iterations,
etc). It turns out that there are 16 of these in CellBlender:

  p1 = Iterations
  p2 = Time Step
  p3 = Maximum Time Step
  p4 = Space Step
  p5 = Interaction Radius
  p6 = Radial Directions
  p7 = Radial Subdivisions
  p8 = Vacancy Search Distance
  p9 = Surface Grid Density
  p10 = Start Seed
  p11 = End Seed
  p12 = Start (viz start)
  p13 = End   (viz end)
  p14 = Step  (viz step)
  p15 = Step  (reaction output step)
  p16 = OutputBufSize

These create a special challenge because they exist when the MCell
property group itself is created.

The "KeyError" was arising from CellBlender code that deleted some of
these when a new data model was replacing the exiting RNA properties.
This happened, for example, during scripting. The "quick fix" in this
commit is to simply not delete them. However, that does end up creating
ID property duplicates in certain circumstances. This is not typically
seen by the user, but dumping the panel parameter ID data may show
multiple copies for some of these core 16 properties. Other than taking
up space, this shouldn't be a problem because the RNA property will
contain a reference to only the most recent version of the ID
properties. So while a dump of the panel parameters might show several
"iterations" values, the CellBlender "iterations" parameter will be set
to the most recent version. Consequently, when the data model is stored,
it is generated only from the single most recent RNA property versions.
So restoring from a data model (when CellBlender is updated, for
example) will remove all of the duplicates.

There should be another fix to this problem that properly resolves the
creation/deletion of these 16 "static" parameters. This fix is intended
to provide a temporary solution until then.