Updated the test suite to use non-Blender icospheres for repeatability.
As with the cube, different Blender versions appear to generate
different mesh orderings when creating an icosphere. This causes a
different MDL file to be written by CellBlender which will not match the
SHA1 checksum of the MDL when CellBlender is run in a different version
of Blender.

As with the cube, the easiest solution was to simply use our own
icosphere generation code. Fortunately, the test suite already had that
code, so it was just a matter of calling the custom version instead of
Blender's version. Also, fortunately, Blender hasn't changed the Torus
generation code yet. This is particularly fortunate since we don't
already have a replacement on hand (although we can do it if needed).

Note that while the cube replacement was identical to Blender's earlier
cube, the icosphere generated by our code isn't numerically identical to
Blender's version. So the MDL checksum (SHA1) for tests using an
icosphere did require a change as reflected in this commit.

This version passes the test suite in both 2.77 and 2.78.