Removed the "type_name" argument from the init_ref function call
The type_name was a required parameter that hadn't been used in some
time. It was preserved during the switch to ID parameters to minimize
the number of things changed at one time.

This commit removes the "type_name" from the parameter system and all
calls that had used it.

This commit has passed the CellBlender test suite.