Stand-alone JSON Data Model to MDL converter
This file is intended to provide a mechanism to export a JSON data model
to MDL. This will be useful for exporting from CellBlender, but it will
also serve other purposes (such as exporting a JSON model from a web
page - or Galaxy - into MDL for MCell to run.

This file is currently identical to io_mesh_mcell_mdl/export_mcell_mdl.py,
but it will be modified to get its data from a data model rather than
CellBlender's properties.