Parameter sweep user interface is functional.
This version contains a relatively complete parameter sweep interface.
It implements most of the recent feature requests except some of the run
limit checks. It can generate a data model containing optional fields
named "sweep_expression" and "sweep_enabled". The sweep_expression is a
string and is not currently checked for any format. The anticipated
format is a comma-separated list of ranges where each range is either a
scalar, a two element start:end sequence or a three element
start:end:step sequence.

This data model is compatible with non-sweeping code because it only
adds new fields without changing the meanings of any existing fields.

The PHP/JavaScript web interface has been updated to read this format
with the limitation that each sweep expression consists of only one
range element of the form start[:end[:step]]. This fits with the current
web page design that contains a start, end, and step for each parameter
being swept.

While the PHP/JavaScript web interface works, CellBlender itself has not
yet been updated to produce MDL based on the sweep parameters. In fact,
the MDL produced will ignore the sweep parameters altogether and operate
only on the original parameter value - even though it is hidden when the
sweep string is being shown. At this point, it's not clear whether the
swept MDL should be produced by CellBlender directly (as it has in the
past), or if it should simply use the new (but currently incomplete)
data_model_to_mdl.py code (in cellblender/mdl). The data_model_to_mdl.py
code is intended to run on its own which makes it handy for use in web
applications (like Galaxy and our PHP/JavaScript demo). But it does
require the additional step of generating a data model before exporting
from Blender.