Added run counting for run limit checking.
Note that there are some potential numerical problems with our
specification of start:stop:step when these are floating point numbers.
Depending on the numbers chosen, a strict interpretation might miss the
end point due to numerical rounding. The algorithm implemented here is
intended to be generous with its inclusion of the ending run. This
should work for cases where the "start:stop" range is intended to be an
integer number of "step" lengths including both end points.