Sweep program "run_data_model_mcell.py" seems to work ... needs testing!
The stand alone program "run_data_model_mcell.py" was adapted from the
"run_simulations.py" program. Code was added to support sweeping. In
particular, it builds a sweep list, creates subdirectories for sweep
points, creates MDL for each sweep point, and runs MCell using the MDL
for each sweep point and in each sweep subdirectory.

This version has a mild inefficiency because it sets the data model
parameters at each seed (which doesn't change the parameters). The next
commit should fix that, but this one was worth saving since it worked.