Implemented sweep running code in CellBlender
The goal for this design is to maintain separate (non-Blender) sweeping
code that can be run outside of CellBlender to support Galaxy and other
web based interfaces. The current version accomplishes this with the two
files in the "mdl" directory:

  data_model_to_mdl.py
  run_data_model_mcell.py

The sweep runner in CellBlender exports a data model and then passes it
to "run_data_model_mcell.py" which uses "data_model_to_mdl.py" to export
the MDL before starting the jobs via Python's multiprocessing module.

The current version will create a separate subdirectory ("sweep_data")
containing additional subdirectories for each parameter being swept. The
ordering of parameters in the tree is the order that they appear in the
data model (CellBlender exports them based on order in the parameter
list). This version exports the data model to "data_model.json" in:

   ProjectName_files/mcell

The JSON format was chosen because that's what web tools (such as
Galaxy and our PHP version) are more likely to support. However, that
might be an option in the future.

This version does not yet create an external "run list" for post
processing. The tree created looks like this:

  Release_Patterns_Sweep_3_no_viz_files
    mcell
      sweep_data
        nr_index_0
          Number_of_Trains_index_0
            react_data
          Number_of_Trains_index_1
            react_data
        nr_index_1
          Number_of_Trains_index_0
            react_data
          Number_of_Trains_index_1
            react_data
        nr_index_2
          Number_of_Trains_index_0
            react_data
          Number_of_Trains_index_1
            react_data

where the parameter "nr" took 3 values and the parameter
"Number_of_Trains" took on two parameters. Note that the number
of seeds produced was not consistent in each run and may need some
further investigation.