Added pre-run creation of react_data & viz_data to avoid race condition
It appears that creating any directory which will be shared by two
or more processes should be done before starting those processes.
In the previous version, the creation of the react_data and viz_data
directories was left to the processes themselves. This caused some
of the seed directories to NOT be created.

This commit creates both a react_data and viz_data directory for each
point in the sweep space whether needed or not. Testing confirms that
this fixes the problem.