Generated a sweep list file named "sweep_list.json"
The data model runner generates a sweep description in the file
"sweep_list.json". The file contains a JSON list of lists:

  [
    ["par_name_1", [v11, v12, v13]],
    ["par_name_2", [v21, v22]],
    ["par_name_3", [v31, v32, v33, v34]],
    ["SEED",       [s1, s2, s3, s4, s5, s6]]
  ]

The last item in the list will be "SEED" which is a reserved word
in MCell and therefore not a legitimate parameter name. It might be
clearest (now that we're naming directories after parameters) to use
the capital "SEED" when naming the seed-related subdirectories so they
cannot be confused with any other parameters. The name "SEED" should
also be an illegal parameter name (although CellBlender doesn't
currently check this).