Sweep plotting updates.
This version will check for the "sweep_data" directory and use it for
plotting if found. Otherwise if will fall back to using the react_data
path.

Right now, only the sweep runner generates the sweep_data directory, and
the sweep_data directory is NOT deleted automatically. So if the sweep
runner is used once and then other (non-sweep) runners are used, the
data in the sweep directory will still be found by the plotting code,
but it will be recognized as stale and no data will be plotted until the
sweep_data directory is removed. Note that the "sweep_data" directory
will be soon changed to "output_data", and all of the runners will
create their output data in that directory. At that time, only legacy
runs will be missing the "output_data" directory, and that will signal
that their data should be expected in the "react_data" directory.