The new output_data layout is working for sweep and command line runners
The sweep runner uses a data model to build the sweep directory tree,
export the MDL, and run mcell.

The command line runner does not use the data model and does not build a
sweep directory tree. Instead, it uses the scalar (non-swept) parameter
values to run a single instance per seed. However, it must handle the
new "output_data" subdirectory (and it does).

Note, that "output_data" may not be the best name since the MDL is also
in that directory tree. While that is "output" from CellBlender, it is
"input" to MCell.