Updated the Blender install directory for the CellBlender libMCell code.
Note that the both the C++ and Pure Python versions worked without any
changes because they don't use MDL. They also don't sweep, but their
output is compatible with the old directory layout. This supports the
hope that this version of CellBlender will be backward compatible with
the old directory layout.