Retain previous viz selections on reload of the same model
The previous version reset all parameter sweep settings to their first
(zero) index when the data was refreshed. This required re-navigating to
the desired sweep position after re-running.

Some work was done to implement a "smart" refresh that would shuffle in
and out the changed values preserving those that hadn't changed. But
that approach hadn't worked, and in the interest of time it was dropped
in favor of the simpler approach of reloading all parameters whenever
the list changed. This commit preserves some of that code (in comments)
in case the approach is desired in the future. It will be removed in
subsequent commits (along with debugging code), so this commit is the
place to find it.