Added support for ranges as list items in a parameter sweep.
A parameter sweep specification can contain comma-separated items where
each item can be either a scalar value, or a range with unit step (b:e),
or a range with arbitrary step (b:e:s).

Note that since these are floating point values, there might be some
ambiguity due to rounding. This code uses a tolerance of step/1000 when
determining when to end the loop. This has the effect of normally
including the end points in expressions like: (3.1:3.7:0.3). This
example will sample at 3.1, 3.4, and 3.7 even in the case where the sum
of (3.1 + 0.3 + 0.3) might exceed 3.7 due to numerical issues and
therefore skip the last point (3.7).