Support Python 2.7 with makedirs_exist_ok function
The previous run_data_model_mcell.py file used Python 3's version of the
makedirs function. That version has an "exist_ok" flag to NOT generate
errors when creating a series of subdirectories where some of them might
already exist. That works fine when this code is run from inside Blender
which uses a Python3.x. However, if this script is run stand-alone in an
enviornment where Python 3 is not available, then it would fail. This
commit adds a function named "makedirs_exist_ok" which is a compatible
replacement for Python 3's version of makedirs (has an "exist_ok" flag).
This will be useful in environments where we cannot guarantee which
version of Python is being used.