Added a "Parameter Sweeping" section to the CellBlender README.md file
CellBlender's Parameter Sweeping is relatively new. This commit provides
a simple specification for the layout of directories and files compatible
with this new capability.

This capability is currently only in the "parameter_sweep" branch of the
code, and it would be good to have it reviewed (and approved) before
merging into the development branch since it creates a new output data
layout that we may feel obligated to support for a long time once it is
in the hands of the CellBlender user community.