Set RAND_UNIFORM and RAND_GAUSSIAN to evaluate to 0 in Parameter System
There is a bit of a problem with the RAND_UNIFORM and RAND_GAUSSIAN MDL
functions when used inside of CellBlender. These functions are used by
MCell at run time to generate random numbers based on a seed. It is not
practical (or even possible?) for CellBlender to calculate appropriate
values for these random number functions to display in the various
parameter fields throughout CellBlender. Even if CellBlender knew the
order that these would be called during the MCell simulation, they would
still depend on the seed which can be multi-valued within CellBlender.

For that reason, this commit will represent both RAND_UNIFORM and
RAND_GAUSSIAN as 0 within the CellBlender user interface.

Note that prior to this change, the two MDL random functions were
directly translated to "uniform" and "gauss" which are the equivalent
python function names. However, that was an error because without
parameters, they both evaluated to methods rather than numbers that
could participate in expressions. In order to have them evaluate
properly, they would have to be substituted as follows:

  RAND_UNIFORM  -> random.uniform(0,1)
  RAND_GAUSSIAN -> random.gauss(0,1)

These substitutions could have been put in get_expression_keywords, but
the varying nature of these functions would still have produced
misleading displays within CellBlender. Even worse, it's not clear if
the paramater system would appropriately handle values that changed
every time they were called. For these reasons, this commit substitutes
the value of '0' for both of those functions.

Note that there is a similar problem with the keyword "SEED" since it
may take on multiple values. It has traditionally been set to 1 within
CellBlender since that is its default value.