Updated CellBlender's "libMCell" to run data model files.
The version of libMCell was started in January of 2016 as a way of
exploring the linkages between CellBlender and a library version of
MCell (libMCell). The project ended up producing skeletons for:

 1. Pure Python version of MCell (pure_python_sim.py)
 2. C++ library (libMCell.cpp) and a C++ main (mcell_main.cpp)
 3. Python main (mcell_main.py) with the C++ library (libMCell.cpp)

Each of these was added as a "runner" option in the development version
of CellBlender. Since the second two (mcell_main.cpp and mcell_main.py)
both use the same libMCell, they should (and currently do) produce the
same output.

Each of these simulations can currently run a CellBlender data model,
however, they only implement a small fraction of that functionality.

These have not been updated based on the recent API explorations.