Application of commit cf803054a42 to the development branch.
Changes were made to the prototype libMCell versions in the
parameter_sweep branch that needed to be brought back into the
development branch. This commit makes those changes without also
implementing the "output_data" directory layer that's being added
in the parameter_sweep branch.

Ideally, these changes should have been made in the development branch
and then merged into parameter_sweep. However, it's tricky because these
changes involved changing code that affected writing of output files
(which is what differs between the two branches). So both branches
needed to be updated. A few more commits might be needed to re-align the
history so future changes to the development branch can continue to be
easily merged into the parameter sweep branch.