Added detection of any reactions to trigger "fake" reactions.
Handling reactions starts by detecting that there are reactions
specified in the model. This commit does the detection and simply
flags that the model "has reactions". That flag is then used to
trigger (or inhibit) the "fake" reactions to remove molecules during
the simulation. There is obviously more work to be done on reactions.