Added the actual sim runner code for Java and OpenGL from top level.
This commit simply duplicates 2 existing sim runners in the new
directory structure. The make files were updated to build them in their
new locations. In the process, the data_plotters makefile was found to
be deficient for cases where there were multiple subdirectories that
needed compilation of some sort. This was not discovered before because
only the Java plotter needed compilation (all others were pure Python).