Rearranging directories to put libMCell testing code into sim_engines
The libMCell testing code is essentially stub code to demonstrate how
user-modified versions of MCell (either C++ or Python) could work with
CellBlender. The libMCell directory had been in the cellblender
directory, but now that we're building simulation engines, it is more
appropriate to put it there. As before, that directory contains a few
different implementations (C++, Python using C++, and Pure Python).
Since they share some common code, they are currently in the same
directory. But to work with the sim engine interface, each simulation
will need its own directory below sim_engine with an __init__.py file.
This will still work with this directory structure by simply referencing
these shared files from the subdirectory for each engine interface. So
the Pure Python version may have its own subdirectory ("pure_python"?)
below sim_engines. However, the __init__.py file in that directory can
simply reference the files from the libMCell directory to run. This is
consistent with the idea that the sim_engine directory need not contain
all of the files to run. It just needs to contain the __init__.py file
that knows where to find what it needs. So if a simulator (xxx_sim) is
installed in some arbitrary directory, the __init__.py file in the
sim_engines directory just needs to know where to find it. This is very
similar to the plotting code that doesn't need to have xmGrace or
GnuPlot directly in the plotting folder. It just needs to be able to
find them. For the plotting code, it searches the entire path since
things like GnuPlot and xmGrace are somewhat likely to already be
installed. For the simulation engines, this kind of search isn't as
useful, but it's up to each __init__.py file to come up with its own
strategy.