Added the pure python engine
Note that because the "engine infrastructure" isn't in place yet, this
different "engine" was added as a different runner. The engine still
lives in the proper place (sim_engines/libMCell/pure_python_sim.py), but
since there's no option to select an engine yet, the easiest way to get
it to run was with a new runner stub in the sim_runners directory.