Added help text for the simulation engines / runners panel.
Previous commits have created a separate panel and added optional
compilation in the makefile to support pluggable engines and runners.
Those changes make this version compatible with the development branch.

This commit adds just a little bit of explanation for anyone finding the
new feature while running the development branch.