Changed "libMCell" folder to "engine_runner_combos"
The history of "engines" and "runners" started with stub alternatives
to mcell implemented as trial versions of "libMCell". Since these were
already being developed, they were easy to integrate into CellBlender as
alternative "runners" (like the Java runner or Queue runner) even though
they were implementing a different simulation algorithm in some cases.

Initially, these lived in the libMCell directory below cellblender, but
were not part of CellBlender itself. As the need to support alternate
runners (for parameter sweeping and multiprocessing) became more clear,
it made sense to include these alternatives as part of CellBlender. They
were initially added to the repository with their original name of
"libMCell", but that is not an accurate name for what they have become.
For that reason (and to avoid confusion), the directory name has been
changed to "engine_runner_combos" which reflects the fact that each one
is both an engine and a runner. They will be maintained in that status
until the implementation of separate engines and runners is complete.