Sun Grid Engine runner functions, but needs more testing.
This is a prototype version that uses a fixed list of machine names to
run. The code actually contains an algorithm to select machines based on
freecpu and capacity (as reported by qhost), but it needs more work. To
get this early version working, the algorithm's results are simply
replaced with nodes 34, 35, 57, and 58.