Added SGE host interface panel (shows when running sweep with SGE)
The panel contains a string property for the SGE scheduler computer
name. Once the name is entered, clicking the refresh button will execute
a qhost command via ssh on the SGE scheduler. The result of the qhost
command is then parsed and converted to a list of computer names,
resources, and a selection button (more bells and whistles may be
added).

This appears to work, however, the resulting list is not yet being
passed to the runner itself (which is run as a separate process). The
current runner still uses several hard-coded "node" computers.

The next step is to define the interface between CellBlender and the SGE
Python runner that supports host names (the current interface supports
sweeping, but the host names are still being forced to a hard-coded list
inside run_data_model_mcell.py). It should also pass the name of the SGE
scheduler (also currently hard coded inside run_data_model_mcell.py).