Pass selected computer names on to "run_data_model_mcell.py"
This version passes the user-selected computer names on to
run_data_model_mcell.py as command line options. Those are
the computer names used when the runner_type option is "sge".

This version appears to work, but there are a number of issues
to be addressed. Most of these are computer resource selection
issues (how to decide which machines to select for running).
The current "run_data_model_mcell.py" has some code that looks
at the "freecpu" parameter and attempts allocation of those
with freecpu > 1.5. But this code is currently over-ridden
by the list of nodes selected in CellBlender.

At this time, it's not clear where that kind of logic belongs.
It should probably be part of the runner interface, since the
runners themselves will contain the intelligence about whatever
run mechanism they're using. But it needs to be presented to
the user by CellBlender. So it might make sense for the runner
interface to have its own ability to draw into a Blender panel.