Removed another hard-coded reference, exploring short qhost return
For some reason, this code produces a short list of machines from the
qhost command. It's not clear where the data is being dropped, and this
commit added some experimental code to try to address the problem. It
remains unresolved in this version, but since the list of machines is
sufficient for most purposes, this is still a functional version.