Fixed the short qhost problem.
For some reason, reading the pipe from the qhost command did not
consistently find a newline. It typically failed on our node39, and so
the list of hosts ended there.

This commit uses different read logic that attempts to read a line even
when the "peek" doesn't show a newline (after a timeout). It's not clear
why this is needed, but it does appear to work and returns all hosts.

This version was tested running 18 jobs, over varying numbers of hosts.
They all seemed to run as expected.