Added host selection based on user-specified memory requirements
Added properties, panels, and logic to support selection of Grid Engine
hosts by a minimum memory requirement.

The current version assumes that memory is listed in Gigabytes with the
"G" suffix in the qhost command.

Also added a timeout when refreshing the host list without having
specified an sge_host_name value. Also added code to check first, but
it's currently commented out in this version to test the timeout.