Cleaned up the memory detection code a bit ... store as float gigabytes.
Note that most of our SGE code operates by parsing the returned text
from SGE commands. The exact specification for the formatting has not
been heavily researched and may not even exist (the printed output being
"obvious" to a human viewer). For example, the qhost command returns:

  hopper      lx24-amd64     12  0.02   35.4G    3.0G   36.0G   1.7M
  hydra       lx24-x86        1  0.17  787.7M  146.6M  589.0M  76.0K

The units of "G", "M", and "K" are assumed to mean multiples of 1024
since that's the common meaning when referring to memory. This code
makes that assumption and also provides for "P" and "E" (although it's
not clear if the SGE was designed to ever produce those units).