Updated code to show total cores and cores in use for SGE Hosts.
There may be a number of features that are useful for selecting
machines. This commit adds the number of free cores to the current
memory selection.

The general procedure is to issue SGE commands, collect results,
incorporate them in the computer database (dictionary) and then
decide which nodes to run based on all of these criteria.