Added "automatic" selection with a memory requirement in SGE runner
It turns out that "automatic" is much easier since there's no need to
deal with host names. Note that while this version shows the free slots
field for both manual and automatic, it's not currently used in the
automatic mode since the grid engine finds the slots. That field will
probably be removed from the automatic version soon.