Replaced the unused free slots with a terminate button in auto mode.
As mentioned in the previous commit, specifying a number of free slots
needed to run doesn't make sense in automatic mode since the Grid Engine
will only assign jobs to computers with free slots. So that field was
replaced with the "Terminate All" operator (which terminates all jobs
owned by the user on the Grid Engine).

This version is complete enough for now. Near term effort will now be
directed at making these runners "pluggable".