Added the generic "Pluggable" code to CellBlender for simulation engines
This code begins to implement a pluggable interface for simulation
engines. It replaces the older code which has been moved to the
"old_sim_engines" directory. It is currently still using the shape
examples until all of the connections can be completed. Then the
actual MCell runner code can be substituted.

The pluggable interface in this version does not yet function, so
continue to use the previously existing interface for running.