Added code to accept either absolute or relative script paths
Blender has a convention that files may be specified as relative to the
.blend file when they start with "//" (at least in Linux). This change
looks for paths that start with that sequence and converts them to an
absolute path for opening the file. This allows paths to be specified as
either absolute or relative. In general, it's better to use "relative"
for files stored in a project (so they will work if the entire project
is moved), and use "absolute" for files shared among projects.