Fixed a minor bug in the molecule visualization sweep panel
The newer sweep code builds an enumerated selector for choosing among
the different values of each swept parameter. It does this by opening a
file named "data_layout.json" in the "output_data" directory in the
"[project]_files" directory. However, if the model hasn't been run, then
some the "[project]_files" directory won't exist and looking for the
"output_data" directory will cause an error. This was mostly a cosmetic
problem (with a nasty Blender pop-up message) since there was nothing to
be done if the file didn't exist anyway. This commit checks that the
"[project]_files" directory exists before trying to look inside of it.