Dynamic runner writes data_layout and Python decodes parameters.
The data layout file could be created by the simulation engines, but
since the format is really for CellBlender's consumption, it makes sense
for CellBlender to create it first. It might be possible for simulation
engines to read from that file as confirmation of what CellBlender is
expecting (maybe in an "after_runs" function).

This commit also updated the limited_python engine to use the parameters
from the data model. The previous version required constants for all
fields. Note that the "_extras" dictionary for each parameter contains a
"par_value" field which should be the numerical value of that parameter.
This makes it easier to translate parameter expressions into numbers
when the expressions are in the general parameters (not in the panel
parameters).

This version passes the CellBlender test suite.