Added basic parameter handling in the C++ prototype.
Many of the CellBlender models being tested with different runners have
parameter values which run fine in the Python prototype but won't run at
all in the C++ prototype. This change adds a simple parameter dictionary
to the C++ prototype to store the parameter name/value pairs found in
the data model and use them when found as values. Note that this quick
implementation assumes that the panel parameters use a single parameter
name (and not a parameter expression). Expressions are fine in the
general parameters (in the parameter panel), but not in other places.