Added output detail controls to prototypes to verify interface
Most of what's needed to run a simulation is contained in the data
model, but there may be other inputs needed by a particular simulator
that aren't in the data model. This example shows an "output detail"
parameter that controls the frequency of printing for the Python and
C++ simulators. The Python simulator also has a "reaction factor"
control that's multiplied by the reaction rate. All of these controls
are useful for demonstrating the interface between the sim engine panels
and the simulators themselves.