Added Cursor Control. Added UV Sphere, Circle, and Grid.
The model objects panel now contains an option to show/hide the cursor
location (uses the cursor icon as a toggle). The hover text is:

  "Show/Hide 3D Cursor Location"

The previous button text on the "Center Cursor" button has been replaced
with an icon. The hover text is:

  "Snap the 3D Cursor to the Center"

This made room to add a few more primitives (circle and grid). The
Monkey was also added, but then removed since it is not a watertight
mesh.

These changes provide greater functionality, and may be helpful in
familiarizing new users with the operation of 3D Cursor.