Added explicit boundaries to the Smoldyn interface
This provides a good example of special parameters needed by a
particular simulator that may not be found in a CellBlender model.

Smoldyn gives an error: "simulation dimensions or boundaries are
undefined" followed by "Simulation skipped" when boundaries are not
included in the model. One approach would be to use CellBlender's
partition controls to define a boundary and then use that from the data
model. While this works, it is less explicit than simply requiring that
boundaries be provided.

This commit adds bounds for x,y,z that can be set individually. It also
adds a "bounding_cube_size" field and a "Set Cube Boundaries" button to
set all dimensions to the same value at once. This implementation
borrows from the "calculator" example to use buttons that change fields.