Removed unused functions to keep the engine as simple as possible.
The build_sweep_list function may be needed in the future, but isn't
used yet, so it was removed. Similarly, the run_simulation (singular)
function is no longer being used, so it was removed. That functionality
is now split between "prepare_runs" (in the engine) and "run_commands"
(in the runner).