Add diffusion limited Lotka-Volterra example
We already had the reaction limited version. This adds in the diffusion
limited version and cleans up the interface a bit. Side note: I ended up
redoing the reaction limited case, so it matches the version in the
chapter instead of the website one. The diffusion limited case in the
chapter had more striking spatial heterogeneities than the version on
the website imo.