Add var rate constant demo. Bugfix loading from DM
This is just a very basic demo of variable rate constants. Halfway
through the simulation the rate constant goes from zero to some non-zero
value causing the molecules in the reaction  to almost entirely
disappear by the end of the simulation.

While adding this demo, I discovered a bug that occurs while loading a
data model with a variable rate constant. The variable rate constant was
stored in the data model correctly, but it couldn't build the properties
from that data model.