Show progress of simulations as percentage.
I added the simulation progress (as a percentage) to each item in the
template list that shows all the current jobs being run. You don't even
have to wiggle the mouse to see if the job is done either, because it's
done using a modal timer. It's currently set to update every 0.1
seconds, but it could be changed to whatever we want. Also, once all the
jobs are done, it will quit checking. It uses a bit of a hack, because
simply updating the template list entry itself won't force a screen
update.

Each entry in the template list now looks something like this:

  PID: 29372, MDL: Scene.main.mdl, Seed: 1, 50%

I'd rather do it as a progress bar, but this will do.

Honestly, I think we could probably do without the MDL section. It's
almost always the same. I'm not sure how useful that is to the user.