Bugfix: race condition and uninitialized value.
To set the percentage finished, we need to parse the stdout for lines
starting with "Iterations". It's possible to get to this block of code
in the modal operator before the stdout has produced a single line
containing the word "Iterations." In that case, last_iter and total_iter
wouldn't be set at all triggering an error.